requirements.txt
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
latex2mathml==3.77.0
loguru==0.7.2
markdown2==2.5.1
minify_html==0.15.0
natsort==8.4.0
pillow==11.0.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
svgwrite==1.4.3
trilium-py==0.9.8
urllib3==2.2.3
wavedrom==2.0.3.post3
